requirements.txt
actionlib==1.14.0
angles==1.9.13
annotated-types==0.7.0
anyio==4.5.2
anytree==2.12.1
apturl==0.5.2
attrs==19.3.0
autobahn==17.10.1
autocommand==2.2.2
Automat==0.8.0
backcall==0.1.0
backports.tarfile==1.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==3.1.1
blinker==1.4
bondpy==1.8.6
breezy==3.0.2
Brlapi==0.7.0
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
catkin-tools==0.9.4
cbor==1.0.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.4.1
click==7.1.2
cmeel==0.57.1
cmeel-assimp==5.4.3.1
cmeel-boost==1.87.0.1
cmeel-console-bridge==1.0.2.2
cmeel-octomap==1.10.0
cmeel-qhull==8.0.2.1
cmeel-tinyxml2==10.0.0
cmeel-urdfdom==4.0.1
cmeel-zlib==1.3.1
coal-library==3.0.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.2
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
dulwich==0.19.15
duplicity==0.8.11.0
dynamic-reconfigure==1.7.3
eigenpy==3.10.3
empy==3.3.2
entrypoints==0.3
exceptiongroup==1.2.2
fastapi>=0.95.0
fasteners==0.14.1
fastimport==0.9.8
Flask==1.1.4
Flask-Cors==3.0.10
future==0.18.2
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
gitdb==4.0.12
GitPython==3.1.44
gpg==1.13.1
greenlet==3.1.1
gTTS==2.5.4
gunicorn==20.1.0
h11==0.14.0
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
igraph==0.11.8
image-geometry==1.16.2
importlib_metadata==8.5.0
importlib_resources==6.4.5
incremental==16.10.1
interactive-markers==1.12.0
ipykernel==5.2.0
ipython==7.13.0
ipython_genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==1.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jedi==0.15.2
Jinja2>=3.1.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-console==6.0.0
jupyter-core==4.6.3
kdl-parser-py==1.14.2
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
laser_geometry==1.6.7
launchpadlib==1.10.10
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
manifold3d==3.0.1
MarkupSafe==1.1.1
matplotlib>=3.7.1
message-filters==1.17.0
mistune==0.8.4
monotonic==1.5
more-itertools==10.5.0
moveit-core==1.1.16
mpi4py==3.0.3
n2g==0.3.3
narwhals==1.28.0
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
networkx==3.1
nose==1.3.7
notebook==6.0.3
numpy>=1.24.2
oauthlib==3.1.0
olefile==0.46
ordered-set==4.1.0
osrf-pycommon==2.0.2
owlready2==0.47
packaging==24.2
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pin==3.4.0
Pint==0.21.1
platformdirs==4.3.6
plotly==6.0.0
probabilistic_model==6.1.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.6.1
psutil==5.5.1
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pybind11==2.13.6
pycairo==1.16.2
pycram_bullet==3.2.8
pycryptodomex==3.6.1
pycups==1.9.73
pydantic>=1.10.7
pydantic_core==2.27.2
pydot==1.4.1
pydub==0.25.1
PyGithub==1.43.7
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==1.9.8
python-dateutil==2.7.3
python-debian===0.1.36build1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-igraph==0.11.8
python-qt-binding==0.4.4
python-snappy==0.5.3
PyTrie==0.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
random-events-lib==0.0.1
random_events==4.1.2
reportlab==3.5.34
requests==2.32.3
requests-unixsocket==0.2.0
resource_retriever==1.12.7
roman==2.0.0
rosbag==1.17.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro==1.0.1
rosdistro-modules==1.0.1
rosgraph==1.17.0
rosinstall==0.7.8
rosinstall-generator==0.1.23
roslaunch==1.17.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.17.0
rosmake==1.15.8
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospkg==1.6.0
rospkg-modules==1.6.0
rospy==1.17.0
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.8
roswtf==1.17.0
rqt-console==0.4.12
rqt-image-view==0.4.17
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rviz==1.14.25
scipy==1.10.1
SecretStorage==2.3.1
Send2Trash==1.5.0
sensor-msgs==1.13.1
service-identity==18.1.0
setuptools-scm==8.2.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==1.9.5
SQLAlchemy==2.0.38
srdfdom==0.6.4
starlette==0.44.0
system-service==0.3
systemd-python==234
tabulate==0.9.0
terminado==0.8.2
testpath==0.4.4
texttable==1.7.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
tomli==2.2.1
topic-tools==1.17.0
tornado==5.1.1
tqdm==4.67.1
traitlets==4.3.3
transforms3d==0.4.2
trimesh==4.6.3
Twisted==18.9.0
txaio==2.10.0
typing_extensions==4.12.2
u-msgpack-python==2.1
ubuntu-advantage-tools==20.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urdfdom-py==0.4.6
urllib3==1.25.8
usb-creator==0.3.7
uvicorn>=0.21.0
vcstool==0.3.0
vcstools==0.1.42
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==2.0.0
wrapt==1.11.2
wsaccel==0.6.2
wstool==0.1.18
xacro==1.14.19
xkit==0.0.0
zipp==3.20.2
zope.interface==4.7.1
python-multipart>=0.0.6
